• Experiment to recurse the the apache install dir for all .conf files.
In response to textmate/apache.tmbundle/#1 the 'Open config..' command now
recurses through the contents of TM_APACHE_DIRECTORY looking for .conf files. Not
yet sure if this is a good idea as the resulting list can be quite large, and in
the case of default installs can include a lot of unused configuration files.